Fixed a weird invocation of crypto/sha1 which was causing a panic.
When "StoreObjectSHA1" was called it was using a weird invocation
of crypto/sha1 by importing crypto and trying to use crypto.SHA1.New
instead of importing crypto/sha1 and using sha1.New. This was causing
a panic with the error "panic: crypto: requested hash function #3 is
unavailable"

This updates the call to use the crypto/sha1 package as documented in
the go docs.